backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
gunicorn==21.2.0
numpy==1.26.3
pandas==2.1.4
networkx==3.2.1
scikit-learn==1.3.2
google-generativeai